+++ lib/install/react.rb
@@ -7,7 +7,7 @@ if File.exist?(babelrc)
   react_babelrc["presets"] ||= []
 
   unless react_babelrc["presets"].include?("react")
-    react_babelrc["presets"].push("react")
+    react_babelrc["presets"].push("@babel/preset-react")
     say "Copying react preset to your .babelrc file"
 
     File.open(babelrc, "w") do |f|
@@ -26,6 +26,6 @@ say "Updating webpack paths to include .jsx file extension"
 insert_into_file Webpacker.config.config_path, "- .jsx\n".indent(4), after: /extensions:\n/
 
 say "Installing all react dependencies"
-run "yarn add react react-dom babel-preset-react prop-types"
+run "yarn add react react-dom @babel/preset-react prop-types"
 
 say "Webpacker now supports react.js 🎉", :green