+++ lib/install/template.rb
@@ -10,6 +10,9 @@ copy_file "#{__dir__}/config/.postcssrc.yml", ".postcssrc.yml"
 say "Copying .babelrc to app root directory"
 copy_file "#{__dir__}/config/.babelrc", ".babelrc"
 
+say "Copying .browserslistrc to app root directory"
+copy_file "#{__dir__}/config/.browserslistrc", ".browserslistrc"
+
 say "Creating JavaScript app source directory"
 directory "#{__dir__}/javascript", Webpacker.config.source_path