+++ package/environments/base.js
@@ -13,6 +13,7 @@ const WebpackAssetsManifest = require('webpack-assets-manifest')
 const CaseSensitivePathsPlugin = require('case-sensitive-paths-webpack-plugin')
 const CompressionPlugin = require('compression-webpack-plugin')
 const UglifyJsPlugin = require('uglifyjs-webpack-plugin')
+const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin')
 
 const { ConfigList, ConfigObject } = require('../config_types')
 const rules = require('../rules')
@@ -45,6 +46,9 @@ const getPluginList = () => {
       publicPath: true
     })
   )
+  if (config.optimize_css) {
+    result.append('OptimizeCSSAssets', new OptimizeCSSAssetsPlugin())
+  }
   if (config.gzip) {
     result.append(
       'Compression',
@@ -92,14 +96,20 @@ const buildUglifier = () => {
     cache: true,
     sourceMap: true,
     uglifyOptions: {
-      ecma: 5,
       compress: {
+        ecma: 5,
         warnings: false,
         comparisons: false
       },
       output: {
+        ecma: 5,
         comments: false,
         ascii_only: true
+      },
+      parse: {
+        // Let uglify-js parse ecma 8 code but always output
+        // ES5 compliant code for older browsers
+        ecma: 8
       },
       safari10: true
     }